requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
cssselect==1.3.0
decorator==4.4.2
Django==5.2
django-crontab==0.7.1
ensta==5.2.9
frozenlist==1.6.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
instagrapi==2.1.3
lxml==5.4.0
moviepy==1.0.3
multidict==6.4.3
neispy==5.0.2
ntplib==0.4.0
numpy==2.2.5
pillow==10.4.0
proglog==0.1.11
propcache==0.3.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pyotp==2.9.0
pyquery==2.0.1
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
